\name{active.erreur}
\alias{active.erreur}
\alias{active.legende}
\alias{active.legende2}
\alias{affine}
\alias{aide}
\alias{barres.moy}
\alias{barres.prop}
\alias{code.ask}
\alias{code.data}
\alias{code.graph}
\alias{code.graph.axes}
\alias{code.graph.barres}
\alias{code.graph.barres.erreurs}
\alias{code.graph.cam}
\alias{code.graph.courbe}
\alias{code.graph.courbe.erreurs1}
\alias{code.graph.courbe.erreurs2}
\alias{code.graph.courbe.leg}
\alias{code.graph.hist}
\alias{code.graph.moust}
\alias{code.graph.nuage}
\alias{code.graph.posleg}
\alias{code.graph.titre}
\alias{code.graphtype}
\alias{code.open}
\alias{col.barres}
\alias{col.bordures}
\alias{col.parts}
\alias{col.symboles}
\alias{convert.variable}
\alias{courbe.moy}
\alias{courbe.prop}
\alias{data.load1}
\alias{data.load2}
\alias{distrib}
\alias{enregistrer}
\alias{fermer.GrapheR}
\alias{fr1.close}
\alias{fr1.openB}
\alias{fr1.openCa}
\alias{fr1.openCo}
\alias{fr1.openD}
\alias{fr1.openH}
\alias{fr1.openM}
\alias{fr1.openN}
\alias{fr2.close}
\alias{fr2.openD}
\alias{fr2.opengraphe}
\alias{fr3.close}
\alias{fr3.openB}
\alias{fr3.openCa}
\alias{fr3.openCoN}
\alias{fr3.openD}
\alias{fr3.openH}
\alias{fr3.openM}
\alias{fr4.close}
\alias{fr4.openB}
\alias{fr4.openCa}
\alias{fr4.openCo}
\alias{fr4.openD}
\alias{fr4.openH}
\alias{fr4.openM}
\alias{fr4.openN}
\alias{fr5.close}
\alias{fr5.openB}
\alias{fr5.openCa}
\alias{fr5.openCo}
\alias{fr5.openH}
\alias{fr5.openM}
\alias{fr5.openN}
\alias{fr6.close}
\alias{fr6.openB}
\alias{fr6.openCo}
\alias{fr6.openM}
\alias{fr7.close}
\alias{graphe.axes}
\alias{graphe.box}
\alias{graphe.erreurs.calculer}
\alias{graphe.erreurs.calculer2}
\alias{graphe.erreurs.tracer}
\alias{graphe.hachures}
\alias{graphe.legende}
\alias{graphe.log}
\alias{graphe.symboles}
\alias{graphe.titre}
\alias{hachures}
\alias{hachures2}
\alias{horizontal}
\alias{language.change}
\alias{language}
\alias{load.GrapheR}
\alias{msg}
\alias{navigation}
\alias{new.window}
\alias{ouvrir.GrapheR}
\alias{pretracer}
\alias{pval}
\alias{reinit.fr7}
\alias{reinit.variables}
\alias{rename.legende}
\alias{rename.legende2}
\alias{rename.noms1}
\alias{rename.nomsparts}
\alias{rename.nomsprop.fac}
\alias{rename.variable}
\alias{symboles}
\alias{texte}
\alias{tracer.barres.limites}
\alias{tracer.barres.moyplusieurs}
\alias{tracer.barres.moyun}
\alias{tracer.barres.propplusieurs}
\alias{tracer.barres.propun}
\alias{tracer.cam}
\alias{tracer.courbe.limites}
\alias{tracer.courbe.moyplusieurs}
\alias{tracer.courbe.moyun}
\alias{tracer.courbe.propplusieurs}
\alias{tracer.courbe.propun}
\alias{tracer.hist.limites}
\alias{tracer.hist}
\alias{tracer.moust}
\alias{tracer.nuage.limites}
\alias{tracer.nuage.plusieurs}
\alias{tracer.nuage.un}
\alias{tracer}
\alias{type.ligne}
\alias{type.trait}
\alias{variables.class}
\alias{vertical}
\title{
GrapheR functions
}
\description{
All intern functions making the interface work.
}
\author{
Maxime Hervé <mx.herve@gmail.com>
}