% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{addConfiguration}
\alias{addConfiguration}
\title{Add configuration}
\usage{
addConfiguration(id, net)
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}

\item{config}{a list with a potentially new node and/or link configuration.}
}
\description{
Add configuration
}
