% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{addEdges}
\alias{addEdges}
\title{Add selected edges.}
\usage{
addEdges(id, v1, v2, color = rgb(0, 0, 0))
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}

\item{v1}{A vector of node ids to connect 'from'.}

\item{v2}{A vector of node ids to connect 'to'.}

\item{color}{A vector of edge colors.}
}
\description{
Add all edges connecting nodes from 'v1' to 'v2' (or vice versa, since graphs are assumed undirected).
}
