% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{center}
\alias{center}
\title{Center the network}
\usage{
center(id)
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}
}
\description{
Centers the grapher network at the center of the canvas' viewport.
}
