% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{change_dimensions}
\alias{change_dimensions}
\title{Change Dimensions}
\usage{
change_dimensions(g, is_3d = FALSE)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{is_3d}{Whether to plot in 3 dimensions or 2 dimensions.}
}
\description{
Change the dimensions of the graph.
}
\examples{
library(shiny)

N <- 500
graph_data <- make_data(N)

ui <- fluidPage(
  radioButtons("dims", "Dimensions", choices = list("3D" = TRUE, "2D" = FALSE)),
  graphOutput("graph")
)

server <- function(input, output) {

  output$graph <- renderGraph({
    graph_data \%>\% 
      graph() \%>\% 
      graph_live_layout(time_step = 5)
  })

  observeEvent(input$dims, {
    graph_proxy("graph") \%>\% 
      change_dimensions(input$dims)
  })

}

\dontrun{shinyApp(ui, server)}

}
