% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{cran_deps_graph}
\alias{cran_deps_graph}
\title{CRAN Dependency Graph}
\usage{
cran_deps_graph(
  max = 10,
  include_base_r = TRUE,
  deps = c("Depends", "Imports", "LinkingTo"),
  format = c("list", "igraph")
)
}
\arguments{
\item{max}{Maximum number of reverse dependencies for a package 
may have.}

\item{include_base_r}{Set to \code{FALSE} to exclude base R packages.}

\item{deps}{Forms of dependencies to take into account.}

\item{format}{The format in which to return the graph, either a 
\code{list} or nodes and edges or an object of class \code{igraph}.}
}
\value{
A list of nodes and links.
}
\description{
Builds the CRAN dependency graph.
}
\examples{
\dontrun{
g <- cran_deps_graph(500, format = "igraph")
min_g <- igraph::mst(g)

min_g \%>\% 
  graph() \%>\% 
  graph_cluster() \%>\% 
  scale_link_color(cluster) \%>\% 
  scale_node_size(degree, c(20, 100))
}

}
