% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{definitions}
\alias{definitions}
\alias{define_nodes_color}
\alias{define_nodes_size}
\alias{define_links_from_color}
\alias{define_links_to_color}
\title{Definitions}
\usage{
define_nodes_color(g, var)

define_nodes_size(g, var)

define_links_from_color(g, var)

define_links_to_color(g, var)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{var}{Bare name of variable to define.}
}
\description{
Define variables to use for color and size.
}
\section{Defaults}{

Nodes:
* `color`
* `size`

Edges
* `fromColor`
* `toColor`
}

\examples{
# generate data
# change custom color
data <- make_data()
data$nodes$myColor <- "#0000ff"

graph(data) \%>\% 
  define_nodes_color(myColor)

}
