% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{enableForce}
\alias{enableForce}
\title{Enables d3-forces.}
\usage{
enableForce(id)
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}
}
\description{
Adds a force-directed layout to the network.
}
