% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{enableLasso}
\alias{enableLasso}
\title{Add lass interaction to the widget.}
\usage{
enableLasso(id)
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}
}
\description{
adds lasso.
}
