% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_graph}
\alias{extract_graph}
\title{Extract Graph}
\usage{
extract_graph(g, json = FALSE)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{json}{= FALSE}
}
\value{
A \code{list} of nodes and edges.
}
\description{
Extract the graph as list, useful if one wants to 
further process it or generate the offline layout
using nodejs as it can be imported with 
\href{ngraph.fromjson}{https://github.com/anvaka/ngraph.fromjson}.
}
\examples{
graph_data <- make_data(10)

g <- graph() \%>\% 
  graph_links(graph_data$links, source, target)

lst <- extract_graph(g)

# json read to import oin ngraph
json <- extract_graph(g, json = TRUE)

}
