% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{make_nodes}
\alias{make_nodes}
\alias{make_links}
\alias{make_data}
\title{Generate data}
\usage{
make_nodes(
  n = 100,
  colors = c("#B1E2A3", "#98D3A5", "#328983", "#1C5C70", "#24C96B")
)

make_links(nodes)

make_data(
  n = 100,
  colors = c("#B1E2A3", "#98D3A5", "#328983", "#1C5C70", "#24C96B")
)
}
\arguments{
\item{n}{Number of nodes.}

\item{colors}{Color palette to use.}

\item{nodes}{Nodes, as generated by \code{sg_make_nodes}.}
}
\value{
\code{tibble} of nodes or links or a \code{list} of the latter.
}
\description{
Generate nodes and links. Forms a hub and spoke graph with
power law distribution.
}
\section{Functions}{

\itemize{
  \item{\code{make_nodes} generate data.frame nodes.}
  \item{\code{make_links} generate data.frame edges.}
  \item{\code{make_data} generate list of nodes and edges.}
}
}

\examples{
graph_data <- make_data()
head(graph_data) # structure

# graph
graph(graph_data)

}
