% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapher.R
\name{graph-shiny}
\alias{graph-shiny}
\alias{graphOutput}
\alias{renderGraph}
\alias{render_graph}
\alias{graphProxy}
\alias{graph_proxy}
\title{Shiny bindings for graph}
\usage{
graphOutput(outputId, width = "100\%", height = "400px")

renderGraph(expr, env = parent.frame(), quoted = FALSE)

render_graph(expr, env = parent.frame(), quoted = FALSE)

graphProxy(id, session = shiny::getDefaultReactiveDomain())

graph_proxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId, id}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a graph}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{session}{A valid shiny session.}
}
\description{
Output and render functions for using graph within Shiny
applications and interactive Rmd documents.
}
