% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{graph_background}
\alias{graph_background}
\title{Background}
\usage{
graph_background(g, color = "#000", alpha = 1)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{color}{Background color of the visualisation.}

\item{alpha}{Transparency of background, a numeric between \code{0} and \code{1}.}
}
\description{
Customise the background color and trasnparency of the visualisation.
}
\examples{
make_data() \%>\% 
  graph() \%>\% 
  graph_background("#d3d3d3", .2)

# as proxy
library(shiny)

graph_data <- make_data(500)

colors <- list(
  "#000000", "#121420",
  "#1B2432", "#2C2B3C"
)

ui <- fluidPage(
  selectInput("bg", "background color", choices = colors),
  graphOutput("graph")
)

server <- function(input, output) {

  output$graph <- renderGraph({
    graph_data \%>\% 
      graph() \%>\% 
      graph_live_layout(time_step = 5)
  })

  observeEvent(input$bg, {
    graph_proxy("graph") \%>\% 
    graph_background(input$bg)
  })

}

\dontrun{shinyApp(ui, server)}

}
