% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{graph_bin_layout}
\alias{graph_bin_layout}
\title{Offline Layout From File}
\usage{
graph_bin_layout(g, positions)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{positions}{Path to binary positions file as computed
by \href{ngraph.offline.layout}{https://github.com/anvaka/ngraph.offline.layout},
generally \code{positions.bin}.}
}
\description{
Add layout computed offline via nodejs. 
Note that \code{\link{graph_offline_layout}} uses the same algorithm.
}
\seealso{
\code{\link{graph_offline_layout}} to run the same algorithm without 
having to export the graph and use nodejs.
}
