% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera.R
\name{graph_camera_fov}
\alias{graph_camera_fov}
\title{Camera FOV}
\usage{
graph_camera_fov(g, fov = 75L)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{fov}{The field of view of the camera.}
}
\description{
Adjust the camera field of view.
}
\examples{
library(shiny)

graph_data <- make_data(500)

colors <- list(
  "#000000", "#121420",
  "#1B2432", "#2C2B3C"
)

ui <- fluidPage(
  sliderInput("fov", "Field of view", 45, 120, 75, step = 5),
  graphOutput("graph")
)

server <- function(input, output) {

  output$graph <- renderGraph({
    graph_data \%>\% 
      graph() \%>\% 
      graph_live_layout(time_step = 5)
  })
r
  observeEvent(input$fov, {
    graph_proxy("graph") \%>\% 
      graph_camera_fov(input$fov)
  })

}

\dontrun{shinyApp(ui, server)}

}
