% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera.R
\name{graph_camera_position}
\alias{graph_camera_position}
\title{Camera Position}
\usage{
graph_camera_position(g, x = NULL, y = NULL, z = NULL)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{x, y, z}{Coordinates of camera position.}
}
\description{
Position the camera, \emph{this only works on a stable graph}:
see \code{\link{graph_stable_layout}} and examples.
}
\examples{
library(shiny)

gdata <- make_data(100)

ui <- fluidPage(
  actionButton("mv", "Move camera"),
  graphOutput("g", height = "80vh")
)

server <- function(input, output) {
  output$g <- render_graph({
    graph(gdata) \%>\% 
      graph_stable_layout(ms = 5000)
  })

  observeEvent(input$mv, {
    graph_proxy("g") \%>\% 
      graph_camera_position(
        sample(1:500, 1),
        sample(1:500, 1),
        sample(1:500, 1)
      )
  })
}

\dontrun{shinyApp(ui, server)}

}
