% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{graph_cluster}
\alias{graph_cluster}
\title{Cluster}
\usage{
graph_cluster(
  g,
  method = igraph::cluster_walktrap,
  quiet = !interactive(),
  ...
)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{method}{The igraph function to create the cluster.}

\item{quiet}{Set to \code{TRUE} to \emph{not print} number 
of clusters in the console.}

\item{...}{Any other bare named column containing 
meta data to attach to the nodes.}
}
\description{
Cluster nodes on the graph. This ultimately adds a \code{cluster}
column to the internal node data.frame. The latter can be used in
e.g.: \code{\link{scale_node_color}}.
}
\examples{
graph_data <- make_data()

graph(graph_data) \%>\% 
  graph_cluster() 

}
