% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{graph_draw}
\alias{graph_draw}
\title{Draw or Redraw a Graph}
\usage{
graph_draw(g)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}
}
\description{
Draw or re-draw a graph.
}
\details{
If used on a \code{\link{graph_proxy}} then
the graph is re-draw, if used on a static visualisation
(\code{\link{graph}}) then the graph is simply drawn. The
latter is only useful if \code{draw} was set to \code{FALSE}
in the initialisation function (\code{\link{graph}}).
}
\examples{
make_data() \%>\% 
  graph(draw = FALSE) \%>\% 
  graph_draw()

}
