% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{graph_events}
\alias{graph_events}
\alias{on_node_click}
\alias{on_node_double_click}
\alias{on_node_hover}
\title{Capture events}
\usage{
on_node_click(g)

on_node_double_click(g)

on_node_hover(g)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}
}
\description{
Enables capturing events in Shiny.
}
\examples{
library(shiny)

graph_data <- make_data(1000)

ui <- fluidPage(
  graphOutput("g", height = "100vh")
)

server <- function(input, output) {
  output$g <- render_graph({
    graph() \%>\% 
      graph_nodes(graph_data$nodes, id) \%>\% 
      graph_links(graph_data$links, source, target) \%>\% 
      on_node_click()
  })

  observeEvent(input$g_node_click, {
    print(input$g_node_click)
  })
}

\dontrun{shinyApp(ui, server)}

}
