% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera.R
\name{graph_focus_node}
\alias{graph_focus_node}
\title{Focus}
\usage{
graph_focus_node(g, id)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{id}{Node id.}
}
\description{
Focus on a specific node.
}
