% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{graph_layout}
\alias{graph_layout}
\title{Layout}
\usage{
graph_layout(
  g,
  spring_length = 30L,
  sping_coeff = 8e-04,
  gravity = -1.2,
  theta = 0.8,
  drag_coeff = 0.02,
  time_step = 20L,
  is_3d = TRUE
)
}
\description{
Adjust graph layout
}
