% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{graph_linklegend}
\alias{graph_linklegend}
\title{Links Legend}
\usage{
graph_linklegend(g, var, color, title = "legend")
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{var}{The bare name of the link variable to use as group.}

\item{color}{The bare name of the link variable to get colors from.}

\item{title}{The title of the legend.}
}
\description{
Add an interactive edge legend using dat.gui.
}
\examples{
data <- make_data()
data$links$grp <- rep(letters[1:2], 5)
data$links$color <- rep(c("#FFFFFF", "#F00FFF"), 5)

graph(data) \%>\% graph_linklegend(grp, color)

}
