% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_links}
\alias{graph_links}
\title{Add Links}
\usage{
graph_links(g, data, source, target, ...)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{data}{A data.frame containing links data.}

\item{source, target}{The bare column names containing 
the links source and target.}

\item{...}{Any other bare named column containing 
meta data to attach to the links.}
}
\description{
Add links to a graph.
}
\examples{
graph_data <- make_data()

graph() \%>\% 
  graph_nodes(graph_data$nodes, id) \%>\% 
  graph_links(graph_data$links, source, target)

# using the shiny proxy
library(shiny)

data1 <- make_data(150)
data2 <- make_data(50)

ui <- fluidPage(
  actionButton("add", "add"),
  graphOutput("g", height = "90vh")
)

server <- function(input, output) {
  output$g <- render_graph({
    graph(data1) 
  })

  observeEvent(input$add, {
    graph_proxy("g") \%>\% 
      graph_links(data2$links, source, target)
  })
}

if(interactive()) shinyApp(ui, server)

}
\seealso{
\code{\link{graph_nodes}} to add nodes.
}
