% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_nodes}
\alias{graph_nodes}
\title{Add Nodes}
\usage{
graph_nodes(g, data, id, ...)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{data}{A data.frame containing nodes data.}

\item{id}{The bare column names containing the nodes ids.}

\item{...}{Any other bare named column containing 
meta data to attach to the nodes.}
}
\description{
Add nodes to the graph.
}
\examples{
graph_data <- make_data()

graph() \%>\% 
  graph_nodes(graph_data$nodes, id)

}
\seealso{
\code{\link{graph_links}} to add links.
}
