% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{graph_stable_layout}
\alias{graph_stable_layout}
\title{Stabilize}
\usage{
graph_stable_layout(g, stable = TRUE)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{stable}{Whether to stabilise the graph.}
}
\description{
Stabilize the layout of the graph.
}
\examples{
library(shiny)

graph_data <- make_data(200)

ui <- fluidPage(
  actionButton("stable", "stabilize"),
  graphOutput("g")
)

server <- function(input, output){
  output$g <- renderGraph({
    graph(graph_data) \%>\% 
      graph_live_layout(time_step = 1)
  })

  gp <- graph_proxy("g")

  observeEvent(input$stable, {
    graph_stable_layout(gp)
  })
}

if(interactive()) shinyApp(ui, server)

}
