% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{graph_static_layout}
\alias{graph_static_layout}
\title{Layout Static}
\usage{
graph_static_layout(
  g,
  method = igraph::layout_nicely,
  dim = 3,
  scaling = NULL,
  ...
)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{method}{The igraph function to compute node positions.}

\item{dim}{Number of dimensions to use, passed to \code{method}.}

\item{scaling}{A vector or 2 values defining the output range to
rescale the coordinates.}

\item{...}{Any other argument to pass to \code{method}.}
}
\description{
Layout the graph given using an igraph algorithm rather 
than the built-in force layout.
}
\note{
This function will overwrite \code{x}, \code{y}, \code{z} variables 
previously passed to \code{\link{graph_nodes}}.
}
\examples{
graph_data <- make_data(10)

g <- graph(graph_data)

# layout
graph_static_layout(g)

# layout with scaling
graph_static_layout(g, scaling = c(-200, 200))

}
