% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapher.R
\name{grapher}
\alias{grapher}
\title{Grapher}
\usage{
grapher(x)
}
\description{
Given a adjacency matrix, or mapper reference object, constructs an interactive grapher htmlwidget.
}
\author{
Matt Piekenbrock
}
