% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{groupSingletons}
\alias{groupSingletons}
\title{Group singletons.}
\usage{
groupSingletons(id, centroid = c(0.9, 0.1), strength = 0.18)
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}

\item{centroid}{an optional 2-vector of XY coordinates to group the singletons around.}
}
\description{
Group singletons.
}
