% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{hex2rgba}
\alias{hex2rgba}
\title{hex2rgba
Converts Hexadecimal colors (with or w/o transparency) to rgba/rgb strings.}
\usage{
hex2rgba(colors)
}
\description{
hex2rgba
Converts Hexadecimal colors (with or w/o transparency) to rgba/rgb strings.
}
