% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{insertLinks}
\alias{insertLinks}
\title{Inserts new links.}
\usage{
insertLinks(id, links, color = rgb(0, 0, 0))
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}

\item{links}{An (n x 2) matrix (edgelist) of links to add.}

\item{color}{A vector of link colors.}
}
\description{
Inserts links connecting nodes given in the \code{links} matrix.
}
