% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{link_distance}
\alias{link_distance}
\alias{hide_distant_links}
\alias{compute_links_distance}
\title{Hide Links}
\usage{
hide_distant_links(g, distance = 1)

compute_links_distance(g)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{distance}{distance above which links should be hidden.}
}
\description{
Hide links over a certain distance, based on computations by \code{graph_static_layout}.
Nodes will actually be hidden in resulting visualisation but not removed.
}
\details{
This is the technique used by \href{Andrei Kashcha}{https://github.com/anvaka},
in his \href{package managers visualisation project}{https://anvaka.github.io/pm}, though
the latter does not use ngraph.pixel (which grapher uses) and hides links based on the 
distance in pixels. Hiding distant edges allows to undo the hairball while still being
able to discern smaller communities.
}
\examples{
gdata <- make_data(500)

graph(gdata) \%>\%
  graph_static_layout(scaling = c(-1000, 1000)) \%>\% 
  graph_cluster() \%>\% 
  scale_link_color(cluster) \%>\%  
  hide_distant_links(150)

}
