% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{logNetwork}
\alias{logNetwork}
\title{Logs the network to the rendering console. Useful for debugging.}
\usage{
logNetwork(id)
}
\description{
Logs the network to the rendering console. Useful for debugging.
}
