% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{pin_nodes}
\alias{pin_nodes}
\alias{pin_node}
\title{Pin Nodes}
\usage{
pin_nodes(g, data, id)

pin_node(g, id)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{data}{A data.frame containing nodes data.}

\item{id}{The bare column names containing the nodes ids.}
}
\description{
Pin nodes in place.
}
