% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove}
\alias{remove}
\alias{graph_drop_nodes}
\alias{graph_drop_node}
\alias{graph_drop_links}
\alias{graph_drop_link}
\title{Remove Nodes & Edges}
\usage{
graph_drop_nodes(g, data, id)

graph_drop_node(g, id)

graph_drop_links(g, data, source, target)

graph_drop_link(g, source, target)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{data}{A data.frame holding nodes id or links source and target.}

\item{id}{Id or bare column name holding node ids to remove.}

\item{source, target}{Bare column names or vectors of length one 
holding source and target defining links to remove.}
}
\description{
Remove nodes and edges from the graph.
}
\examples{
library(shiny)

N <- 250

data <- make_data(N)

ui <- fluidPage(
  actionButton("rm", "remove random nodes"),
  graphOutput("g", height = "90vh")
)

server <- function(input, output) {
  output$g <- render_graph({
    graph(data) 
  })

observeEvent(input$rm, {
  graph_proxy("g") \%>\% 
    graph_drop_nodes(
      dplyr::slice(data$nodes, sample(1:N, 2)), 
      id
    ) 
  })
}

if(interactive()) shinyApp(ui, server)

}
