% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{removeLinks}
\alias{removeLinks}
\title{Remove selected links.}
\usage{
removeLinks(id, links)
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}

\item{links}{An (n x 2) matrix (edgelist) of links to remove.}
}
\description{
Each entry in the 'links' matrix is assumed to be a node id. Order does not matter, i.e. 
\code{removeLinks(id, cbind(0, 1))} is equivalent to \code{removeLinks(..., cbind(1, 0))}.
}
