% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{removeNodes}
\alias{removeNodes}
\title{#' Generates the JSON configuration of a given grapher object. 
#' @param id Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in. 
#' @export
get_json_config <- function(id){
  # if (!is(g, "grapher")){ stop("'get_json_config' expects a grapher object.") }
  export <- list(id = id, method = "getGraph")
  callJS()
}
Remove selected nodes.}
\usage{
removeNodes(id, node_ids)
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}

\item{node_ids}{A vector of node ids to remove.}
}
\description{
#' Generates the JSON configuration of a given grapher object. 
#' @param id Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in. 
#' @export
get_json_config <- function(id){
  # if (!is(g, "grapher")){ stop("'get_json_config' expects a grapher object.") }
  export <- list(id = id, method = "getGraph")
  callJS()
}
Remove selected nodes.
}
