% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{rescale_layout}
\alias{rescale_layout}
\title{Rescale Layout}
\usage{
rescale_layout(g, scale = c(-1000, 1000))
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{scale}{A vector or 2 values defining the output range to
rescale the coordinates.}
}
\description{
Rescale the coordinates of the layout.
}
\examples{
g <- make_data()

graph(g) \%>\% 
  graph_layout_static(scaling = NULL) \%>\% 
  rescale_layout()

}
