% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve}
\alias{retrieve}
\alias{retrieve_node}
\alias{retrieve_link}
\title{Retrieve}
\usage{
retrieve_node(g, id)

retrieve_link(g, source, target)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{id}{Id of the node to retrieve.}

\item{source, target}{Source and target id of link.}
}
\description{
Retrieve a node's data.
}
\examples{
library(shiny)

gdata <- make_data(100)

ui <- fluidPage(
 sliderInput("node", "retrieve", 1, max = 100, value = 5),
 graphOutput("g", height = "50vh"),
 verbatimTextOutput("selected")
)

server <- function(input, output) {
 output$g <- render_graph({
   graph(gdata) \%>\% 
   graph_stable_layout(ms = 1500)
 })

 observeEvent(input$node, {
   graph_proxy("g") \%>\% 
     retrieve_node(input$node)
 })

 output$selected <- renderPrint({
   print(input$g_retrieve_node)
 })
}

if(interactive()) shinyApp(ui, server)

}
