% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_link_color}
\alias{scale_link_color}
\alias{scale_link_source_color}
\alias{scale_link_target_color}
\title{Scale Edge Color}
\usage{
scale_link_source_color(g, variable, palette = graph_palette())

scale_link_target_color(g, variable, palette = graph_palette())

scale_link_color(g, variable, palette = graph_palette())
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{variable}{Bare column name of variable to scale against.}

\item{palette}{Color palette.}
}
\description{
Scale nodes color.
}
