% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_node_color}
\alias{scale_node_color}
\title{Scale Node Color}
\usage{
scale_node_color(g, variable, palette = graph_palette())
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{variable}{Bare column name of variable to scale against.}

\item{palette}{Color palette.}
}
\description{
Scale nodes color.
}
\examples{
graph_data <- make_data(100)
graph_data$nodes$var <- runif(100, 1, 10)

graph() \%>\% 
  graph_nodes(graph_data$nodes, id, var) \%>\% 
  graph_links(graph_data$links, source, target) \%>\% 
  scale_node_color(var)

}
