% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_node_size}
\alias{scale_node_size}
\title{Scale Node Size}
\usage{
scale_node_size(g, variable, range = c(20, 70))
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{variable}{Bare column name of variable to scale against.}

\item{range}{Output range.}
}
\description{
Scale nodes size.
}
\examples{
graph_data <- make_data()

graph_data \%>\% 
  graph() \%>\% 
  scale_node_size(size)

}
