% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{setLinkColor}
\alias{setLinkColor}
\title{Sets the link colors.}
\usage{
setLinkColor(id, color)
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}

\item{color}{A vector of hexadecimal color codes. Transparency values are *not* supported.}
}
\description{
Sets the link colors.
}
