% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{setNodeColor}
\alias{setNodeColor}
\title{Sets the node colors.}
\usage{
setNodeColor(id, color)
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}

\item{color}{A vector of hexadecimal color codes. Transparency values supported.}
}
\description{
Sets the node colors.
}
