% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{setNodes}
\alias{setNodes}
\title{Set specific node properties.}
\usage{
setNodes(id, node_ids, node_config)
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}

\item{node_ids}{A vector of node ids to update.}

\item{node_config}{A json configuration of node properties.}
}
\description{
Set specific node properties.
}
