\name{Swallows}
\docType{data}
\alias{Swallows}
\title{Swallows dataset}
\description{An illustrative and fictive dataset about African and European swallows.}
\usage{Swallows}
\format{
\enumerate{
  \item Species
  \item Sex
  \item Size
  \item Weight
  \item Coconuts
  \item Distance
}
}
\keyword{datasets}