% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{toggleLasso}
\alias{toggleLasso}
\title{Disables lass interaction to the widget.}
\usage{
toggleLasso(id)
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}
}
\description{
disables lasso.
}
