% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_link}
\alias{update_link}
\alias{update_link_source_color}
\alias{update_link_target_color}
\title{Update Link}
\usage{
update_link_source_color(g, source, target, val)

update_link_target_color(g, source, target, val)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{source, target}{Source and Target ids of link to update.}

\item{val}{The updated value to assign.}
}
\description{
Update a link source, and target color.
}
