% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_node}
\alias{update_node}
\alias{update_node_size}
\alias{update_node_color}
\title{Update Node}
\usage{
update_node_size(g, id, var, val)

update_node_color(g, id, var, val)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{id}{Id of node to update.}

\item{var}{Variable name to update.}

\item{val}{New value to assign to \code{var}.}
}
\description{
Update a node's color or size, works with /code{\link{graph_proxy}}.
}
\examples{
library(shiny)

N <- 200
g <- make_data(N)

ui <- fluidPage(
 numericInput("node", "Node ID", 5, min = 1, max = N, step = 1L),
 selectInput("color", "New Color", c("red", "green", "blue")),
 graphOutput("g", height = "80vh")
)

server <- function(input, output) {
 output$g <- render_graph({
   graph(g)
 })

 observeEvent(c(input$color, input$node), {
   graph_proxy("g") \%>\% 
     update_node_color(input$node, "color", input$color)
 })
}

if(interactive()) shinyApp(ui, server)

}
