% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{updateGrapherNodes}
\alias{updateGrapherNodes}
\title{Update specific nodes.}
\usage{
updateGrapherNodes(id, node_ids, net_config)
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}

\item{node_ids}{A vector of node ids to update.}

\item{net_config}{A json}
}
\description{
Update specific nodes.
}
