% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{updateNodeSize}
\alias{updateNodeSize}
\title{Update the node sizes.}
\usage{
updateNodeSize(id, r)
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}

\item{r}{A vector of integers representing the node's relative size.}
}
\description{
Update the node sizes.
}
