\name{All GrapheR functions}
\alias{affine}
\alias{ajouter.droite1}
\alias{ajouter.droite2}
\alias{back}
\alias{barres}
\alias{barres.col.barres}
\alias{barres.col.bordures}
\alias{barres.erreur.act}
\alias{barres.erreur.types}
\alias{barres.erreurs}
\alias{barres.fact1}
\alias{barres.fact2}
\alias{barres.legende}
\alias{barres.motifs}
\alias{barres.moy1}
\alias{barres.moy2}
\alias{barres.prop1}
\alias{barres.prop2A}
\alias{barres.prop2B}
\alias{barres.som1}
\alias{barres.som2}
\alias{barres.stack}
\alias{barres.type}
\alias{barres.type1}
\alias{barres.type2}
\alias{barres.type3}
\alias{barres.var}
\alias{barres.yinf}
\alias{barres.ysup}
\alias{calc.err1}
\alias{calc.err2}
\alias{calc.err3}
\alias{calc.err4}
\alias{calc.err5}
\alias{cam.fact}
\alias{cam.num}
\alias{camembert}
\alias{camembert.var}
\alias{charger}
\alias{commencer}
\alias{couleur}
\alias{courbe}
\alias{courbe.droite}
\alias{courbe.erreur.act}
\alias{courbe.erreur.types}
\alias{courbe.fact}
\alias{courbe.legende}
\alias{courbe.symbol}
\alias{courbe.type1}
\alias{courbe.type2}
\alias{courbe.var2}
\alias{courbe.vary}
\alias{courbes.axes}
\alias{courbes.erreurs1}
\alias{courbes.erreurs2}
\alias{courbes.moy1}
\alias{courbes.moy2}
\alias{courbes.moy3}
\alias{courbes.prop1}
\alias{courbes.prop2}
\alias{courbes.prop3}
\alias{courbes.som1}
\alias{courbes.som2}
\alias{courbes.som3}
\alias{distrib}
\alias{graphe.choix}
\alias{histogram}
\alias{histogram.distrib}
\alias{histogram.fact}
\alias{histogram.type1}
\alias{histogram.type2}
\alias{histogram.type3}
\alias{horizontal}
\alias{legende}
\alias{moustaches}
\alias{moustaches.fact}
\alias{new.window}
\alias{niv}
\alias{noms.parts}
\alias{noms1}
\alias{nuage}
\alias{nuage.droite}
\alias{nuage.fact}
\alias{nuage.legende}
\alias{nuage.symbol}
\alias{nuage1}
\alias{nuage2}
\alias{parts}
\alias{plusieurscouleurs}
\alias{plusieurscourbes}
\alias{plusieursdroites}
\alias{plusieursmotifs}
\alias{plusieurssymboles}
\alias{prop.barres}
\alias{pvalue}
\alias{renommer}
\alias{sauver}
\alias{texte}
\alias{toolbar}
\alias{tracer.barres}
\alias{tracer.camembert}
\alias{tracer.courbe}
\alias{tracer.hist}
\alias{tracer.hist.effdens}
\alias{tracer.hist.freq}
\alias{tracer.legende.1}
\alias{tracer.legende.2}
\alias{tracer.legende.3}
\alias{tracer.moustaches}
\alias{tracer.nuage}
\alias{transformer}
\alias{unecouleur}
\alias{unecourbe}
\alias{unedroite}
\alias{unmotif}
\alias{unsymbole}
\alias{vertical}
\title{
GrapheR functions
}
\description{
All intern functions making the interface work.
}
\author{
Maxime Hervé <maxime.herve@laposte.net>
}