% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{definitions}
\alias{definitions}
\alias{define_node_color}
\alias{define_node_size}
\alias{define_link_source_color}
\alias{define_link_target_color}
\alias{define_link_hidden}
\title{Definitions}
\usage{
define_node_color(g, var)

define_node_size(g, var)

define_link_source_color(g, var)

define_link_target_color(g, var)

define_link_hidden(g, var)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{var}{Bare name of variable to define.}
}
\description{
Define variables to use for color, size, and hidden links.
}
\section{Defaults}{

Nodes:
* `color`
* `size`

Edges
* `fromColor`
* `toColor`
}

\examples{
# generate data
# add custom color
data <- make_data()
data$nodes$myColor <- "#0000ff"

graph(data) \%>\% 
  define_node_color(myColor)

}
