% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapher.R
\name{graph}
\alias{graph}
\title{Initialise}
\usage{
graph(width = "100\%", height = NULL, elementId = NULL)
}
\arguments{
\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Id of element.}
}
\description{
Initialise a grapher graph.
}
