% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapher.R
\name{graph}
\alias{graph}
\title{Initialise}
\usage{
graph(
  data = NULL,
  directed = TRUE,
  draw = TRUE,
  width = "100\%",
  height = "100vh",
  elementId = NULL
)
}
\arguments{
\item{data}{A named \code{list} containing \code{nodes} and \code{links} 
where the former's first column are the node ids and the latter's first
and second columns are source and target, every other column is added as
respective meta-data. Can also be an object of class \code{igraph} from the
\link[igraph]{igraph} package. If a character string is passed the string is
assumed to be the path to either a \code{.gexf} file or a \code{.gv} 
\href{https://en.wikipedia.org/wiki/DOT_(graph_description_language)}{dot file}.
If \code{NULL} data \emph{must} be later supplied with \code{\link{graph_nodes}}
and \code{\link{graph_links}}.}

\item{directed}{Whether the graph is directed, if passing an object of class 
\code{igraph} to \code{data} then this is inferred from the object.}

\item{draw}{If \code{FALSE} the graph is not rendered.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Id of element.}
}
\description{
Initialise a grapher graph.
}
\examples{
g <- make_data(50) # mock data

# from a list
graph(g)

# from igraph
ig <- igraph::make_ring(10)
graph(ig)

# from tidygraph
tbl_graph <- tidygraph::create_ring(20)
graph(tbl_graph)

# from gexf
graph("http://gephi.org/datasets/LesMiserables.gexf")

# from dot file
fl <- system.file("example/dotfile.gv", package = "grapher")
graph(fl)

# from data.frames
# pass only links
graph() \%>\% 
  graph_links(g$links, source, target)

# pass nodes and links
graph() \%>\% 
  graph_nodes(g$nodes, id) \%>\% 
  graph_links(g$links, source, target)

}
