% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_links}
\alias{graph_links}
\title{Add Edges}
\usage{
graph_links(g, data, source, target, ...)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}}.}

\item{data}{A data.frame containing edges data.}

\item{source, target}{The bare column names containing 
the edges source and target.}

\item{...}{Any other bare named column containing 
meta data to attach to the edges.}
}
\description{
Add Edges
}
\examples{
graph_data <- make_data()

graph() \%>\% 
  graph_nodes(graph_data$nodes, id, color = color) \%>\% 
  graph_links(graph_data$links, source, target)

}
\seealso{
\code{\link{graph_nodes}} to add nodes.
}
