\name{GrapheR-package}
\alias{GrapheR-package}
\alias{GrapheR}
\docType{package}
\title{
A Multi-Platform GUI for Drawing Customizable Graphs in R
}
\description{
A multi-platform user interface for drawing highly customizable graphs in R. It aims to be a valuable help to quickly draw publishable graphs without any knowledge of R commands. Six kinds of graph are available: histogram, box-and-whisker plot, bar plot, pie chart, curve and scatter plot.
}
\details{
\tabular{ll}{
Package: \tab GrapheR\cr
Type: \tab Package\cr
Version: \tab 1.9-86-5\cr
Date: \tab 2023-11-06\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Maxime HERVE

Maintainer: Maxime HERVE <maxime.herve@univ-rennes1.fr>
}

\references{
    DALGAARD P (2001) A Primer on the R-Tcl/Tk Package.
    \emph{R News}, \bold{1(3)}, 27--31.

    DALGAARD P (2002) Changes to the R-Tcl/Tk Package.
    \emph{R News}, \bold{2(3)}, 25--27.

    HERVE M (2011) GrapheR: a multiplatform GUI for drawing customizable graphs in R.
    \emph{The R Journal}, \bold{3(2)}, 45--53.
}
\examples{
\dontrun{
# To launch or re-launch the interface after closing:
run.GrapheR()
}
}