% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove}
\alias{remove}
\alias{graph_nodes_remove}
\alias{graph_links_remove}
\title{Remove Nodes & Edges}
\usage{
graph_nodes_remove(g, data, id)

graph_links_remove(g, data, source, target)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{data}{A data.frame holding nodes id or links source and target.}

\item{id}{Bare column name holding node ids to remove.}

\item{source, target}{Bare column names holding source and target defining
links to remove.}
}
\description{
Remove nodes and edges from the graph.
}
\examples{
library(shiny)

N <- 250

data <- make_data(N)

ui <- fluidPage(
  actionButton("rm", "remove random nodes"),
  graphOutput("g", height = "90vh")
)

server <- function(input, output) {
  output$g <- render_graph({
    graph(data) 
  })

observeEvent(input$rm, {
  graph_proxy("g") \%>\% 
    graph_nodes_remove(
      dplyr::slice(data$nodes, sample(1:N, 2)), 
      id
    ) 
  })
}

if(interactive()) shinyApp(ui, server)

}
