% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{remove_coordinates}
\alias{remove_coordinates}
\title{Remove Coordinates}
\usage{
remove_coordinates(g)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}
}
\description{
Removes coordinates (\code{x}, \code{y}, and \code{z}) from the 
graph. This is useful if you have used the \code{graph_static_layout}
to then \code{hide_long_links} but nonetheless want to use the default 
force layout (\code{graph_live_layout}).
}
\examples{
gdata <- make_data(500)

graph(gdata) \%>\%
  graph_static_layout() \%>\% 
  hide_long_links(20) \%>\% # hide links
  remove_coordinates() 

}
