% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{removeNodes}
\alias{removeNodes}
\title{Remove selected nodes.}
\usage{
removeNodes(id, node_ids)
}
\arguments{
\item{id}{Either the grapher's htmlwidget instance or, if in a shiny context, the container id the widget is housed in.}

\item{node_ids}{A vector of node ids to remove.}
}
\description{
Remove selected nodes.
}
