\name{run.GrapheR}
\alias{run.GrapheR}
\title{
Run GrapheR
}
\description{
The function launches GrapheR. Use it after closing the interface to re-launch it.
}
\usage{
run.GrapheR(lang=NULL)
}
\arguments{
  \item{lang}{
language, when interface relaunching.
}
}
\author{
Maxime Hervé <mx.herve@gmail.com>
}