% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{save_graph}
\alias{save_graph}
\title{Save}
\usage{
save_graph(g, file, ...)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{file}{Name of file to save graph, passed to 
\link[htmlwidgets]{saveWidget}.}

\item{...}{Other arguments passed to
\link[htmlwidgets]{saveWidget}.}
}
\description{
Save the graph as HTML file.
}
\examples{
\dontrun{
make_data(10)
 graph(data) \%>\% 
 save_graph("grapher.html")
}

}
