% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_node_color}
\alias{scale_node_color}
\alias{scale_node_color_coords}
\title{Scale Node Color}
\usage{
scale_node_color(g, variable, palette = graph_palette())

scale_node_color_coords(g, red = c(0.01, 0.99), green = red, blue = red)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{variable}{Bare column name of variable to scale against.}

\item{palette}{Color palette.}

\item{red, green, blue}{The possible range of values (light) that the
red, green, and blue channels can take, must be vectors ranging from
\code{0} to \code{1}.}
}
\description{
Scale nodes color.
}
\examples{
graph_data <- make_data(100)
graph_data$nodes$var <- runif(100, 1, 10)

# scale by variable
graph() \%>\% 
  graph_nodes(graph_data$nodes, id, var) \%>\% 
  graph_links(graph_data$links, source, target) \%>\% 
  scale_node_color(var)

# scale by coordinate position
graph(graph_data) \%>\% 
 graph_static_layout() \%>\% 
 scale_node_color_coords()

}
