% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_link}
\alias{update_link}
\alias{update_link_source_color}
\alias{update_link_target_color}
\title{Update Link}
\usage{
update_link_source_color(g, source, target, val)

update_link_target_color(g, source, target, val)
}
\arguments{
\item{g}{An object of class \code{graph} as 
returned by \code{\link{graph}} or a \code{graph_proxy}
as returned by a function of the same name.}

\item{source, target}{Source and Target ids of link to update.}

\item{val}{The updated value to assign.}
}
\description{
Update a link source, and target color.
}
\examples{
library(shiny)

N <- 200
g <- make_data(N)

ui <- fluidPage(
 numericInput("link", "link", 5, min = 1, max = N, step = 1L),
 graphOutput("g", height = "80vh")
)

server <- function(input, output) {
 output$g <- render_graph({
   graph(g)
 })

 observeEvent(input$link, {
   sel <- dplyr::slice(g$links, input$link)
   graph_proxy("g") \%>\% 
     update_link_source_color(sel$source, sel$target, "blue") \%>\% 
     update_link_target_color(sel$source, sel$target, "red")
 })
}

if(interactive()) shinyApp(ui, server)

}
